Successful network request and getting a response

+++ app/src/main/java/com/example/topmama/domain/models/Current.kt
@@ -6,7 +6,7 @@ data class Current(
     val humidity: Int,
     val is_day: String,
     val observation_time: String,
-    val precip: Int,
+    val precip: Float,
     val pressure: Int,
     val temperature: Int,
     val uv_index: Int,
